Moves build steps to package.json and splits them
This lets me watch specific directories and invoke different build scripts for
each one, instead of the monolithic build script, which was taking longer than
I expected it to.

After doing this, what I realized is that watching might be the long pole in the
tent instead of the build script.  Next step, try a different watcher.